Converted everything to i32

+++ wasm/graphics/lcd.ts
@@ -38,7 +38,7 @@ export function setLcdStatus(lcdEnabledStatus: boolean): void {
   // 2 or 10: Searching Sprites Atts
   // 3 or 11: Transfering Data to LCD Driver
 
-  let lcdStatus: u8 = eightBitLoadFromGBMemory(Graphics.memoryLocationLcdStatus);
+  let lcdStatus: i32 = eightBitLoadFromGBMemory(Graphics.memoryLocationLcdStatus);
   if(!lcdEnabledStatus) {
     // Reset scanline cycle counter
     Graphics.scanlineCycleCounter = 0;
@@ -56,10 +56,10 @@ export function setLcdStatus(lcdEnabledStatus: boolean): void {
   }
 
   // Get our current scanline, and lcd mode
-  let scanlineRegister: u8 = eightBitLoadFromGBMemory(Graphics.memoryLocationScanlineRegister);
-  let lcdMode: u8 = lcdStatus & 0x03;
+  let scanlineRegister: i32 = eightBitLoadFromGBMemory(Graphics.memoryLocationScanlineRegister);
+  let lcdMode: i32 = lcdStatus & 0x03;
 
-  let newLcdMode: u8 = 0;
+  let newLcdMode: i32 = 0;
   let shouldRequestInterrupt: boolean = false;
 
   // Find our newLcd mode
@@ -109,7 +109,7 @@ export function setLcdStatus(lcdEnabledStatus: boolean): void {
 
     // Check for the coincidence flag
     // Need to check on every mode, and not just HBLANK, as checking on hblank breaks shantae, which checks on vblank
-    let coincidenceCompare: u8 = eightBitLoadFromGBMemory(Graphics.memoryLocationCoincidenceCompare);
+    let coincidenceCompare: i32 = eightBitLoadFromGBMemory(Graphics.memoryLocationCoincidenceCompare);
     if((newLcdMode === 0 || newLcdMode === 1) &&
       scanlineRegister === coincidenceCompare) {
       lcdStatus = setBitOnByte(2, lcdStatus);
